fixing line format and code structure
