+++ src/qutip_qip/algorithms/qft.py
@@ -9,7 +9,7 @@ from qutip import Qobj
 from ..decompose import decompose_one_qubit_gate
 
 
-__all__ = ["qft", "qft_steps", "qft_gate_sequence"]
+__all__ = ["qft", "qft_steps", "qft_gate_sequence", "inverse_qft_gate_sequence"]
 
 
 def qft(N=1):
@@ -133,6 +133,56 @@ def qft_gate_sequence(N=1, swapping=True, to_cnot=False):
                 qc.add_gate("SWAP", targets=[N - i - 1, i])
     return qc
 
+def inverse_qft_gate_sequence(N=1, swapping=True, to_cnot=False):
+    """
+    Create the inverse quantum Fourier transform gate sequence.
+    
+    Parameters
+    ----------
+    N : int, optional
+        Number of qubits (default: 1)
+    swapping : bool, optional
+        Flag to include SWAP gates at the beginning (default: True)
+    to_cnot : bool, optional
+        Flag to decompose controlled phase gates to CNOT gates (default: False)
+        
+    Returns
+    -------
+    QubitCircuit
+        Gate sequence implementing inverse QFT
+    """
+    if N < 1:
+        raise ValueError("Minimum value of N can be 1")
+    
+    qc = QubitCircuit(N)
+    
+    if N == 1:
+        qc.add_gate("SNOT", targets=[0])
+    else:
+        if swapping:
+            for i in range(N // 2):
+                qc.add_gate("SWAP", targets=[N - i - 1, i])
+        
+        # Inverse QFT main algorithm
+        for i in range(N - 1, -1, -1):
+            qc.add_gate("SNOT", targets=[i])
+            
+            for j in range(i - 1, -1, -1):
+                if not to_cnot:
+                    qc.add_gate(
+                        "CPHASE",
+                        targets=[j],
+                        controls=[i],
+                        arg_label=r"{-\pi/2^{%d}}" % (i - j),
+                        arg_value=-np.pi / (2 ** (i - j)),
+                    )
+                else:
+                    decomposed_gates = _cphase_to_cnot(
+                        [j], [i], -np.pi / (2 ** (i - j))
+                    )
+                    qc.gates.extend(decomposed_gates)
+    
+    return qc
 
 def _cphase_to_cnot(targets, controls, arg_value):
     rotation = Qobj([[1.0, 0.0], [0.0, np.exp(1.0j * arg_value)]])
